requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1

langchain==0.3.13
langchain-openai==0.2.14
langchain-groq==0.2.1
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
langchain-community==0.3.13

chromadb==0.6.0
sentence-transformers==3.3.1
torch>=2.1.0

pydantic==2.10.4